Isolate UpdateClient methods
