+++ crates/relayer-cosmos/src/lib.rs
@@ -9,5 +9,6 @@ extern crate alloc;
 pub mod all_for_one;
 pub mod contexts;
 pub mod impls;
+pub mod methods;
 pub mod traits;
 pub mod types;
