+++ crates/relayer-cosmos/src/methods/mod.rs
@@ -1,0 +1,1 @@
+pub mod update_client;
